chore: types.d.ts 정리
