--- src/type.ts
@@ -1,5 +1,0 @@
-export interface Writer {
-  image: string | null;
-  nickname: string;
-  id: number;
-}
