--- src/types.ts
@@ -1,43 +1,0 @@
-import { StaticImageData } from "next/image";
-
-export interface Writer {
-    image: string;
-    nickname: string;
-    id: number;
-}
-
-export interface Comment {
-    epigramId: number;
-    writer: Writer;
-    updatedAt: string;
-    createdAt: string;
-    isPrivate: boolean;
-    content: string;
-    id: number;
-}
-
-export interface CommentsResponse {
-    totalCount: number;
-    nextCursor: number;
-    list: Comment[];
-}
-
-
-export interface EmotionData {
-    emotion: string;
-    rate: number;
-    image: StaticImageData;
-    label: string;
-    color: string;
-}
-
-export interface EmotionChartData extends EmotionData {
-    deg: number;
-}
-
-export interface MonthlyEmotionResponse {
-    id: number,
-    userId: number,
-    emotion: string,
-    createdAt: Date,
-}
